refactor(workflow): Move getParameterValue to WorkflowExpression
Move all parameter resolution logic from Expression.getParameterValue
to WorkflowExpression.getParameterValue to improve performance and fix
timezone handling.

Changes:
- WorkflowExpression.getParameterValue now handles full recursive
  parameter resolution (arrays, objects, expressions)
- Added early return for simple non-expression values to avoid
  unnecessary WorkflowDataProxy creation
- WorkflowDataProxy created AFTER early returns (not before)
- Expression.getParameterValue removed - Expression now only has
  resolveSimpleParameterValue and convertObjectValueToString
- Updated Expression tests to use resolveSimpleParameterValue directly
- Use workflow.timezone directly instead of re-extracting from settings

Impact:
- Fixes Gmail node test timezone bug (prepareTimestamp now runs with
  correct SystemZone before WorkflowDataProxy sets timezone)
- Performance improvement: WorkflowDataProxy not created for literal
  parameter values like "message", "getAll", etc.
- Expression class further simplified - only expression evaluation logic
