+++ packages/workflow/src/expression.ts
@@ -453,57 +453,4 @@ export class Expression {
 	 * @param {IWorkflowDataProxyData} data - The workflow data proxy data
 	 * @param {boolean} [returnObjectAsString=false] - Whether to convert objects to strings
 	 */
-	getParameterValue(
-		parameterValue: NodeParameterValueType | INodeParameterResourceLocator,
-		data: IWorkflowDataProxyData,
-		returnObjectAsString = false,
-	): NodeParameterValueType {
-		// Helper function which returns true when the parameter is a complex one or array
-		const isComplexParameter = (value: NodeParameterValueType) => {
-			return typeof value === 'object';
-		};
-
-		// Check if it value is a simple one that we can get it resolved directly
-		if (!isComplexParameter(parameterValue)) {
-			return this.resolveSimpleParameterValue(
-				parameterValue as NodeParameterValue,
-				data,
-				returnObjectAsString,
-			);
-		}
-
-		// The parameter value is complex so resolve depending on type
-		if (Array.isArray(parameterValue)) {
-			// Data is an array
-			const returnData = parameterValue.map((item) =>
-				this.getParameterValue(item as NodeParameterValueType, data, returnObjectAsString),
-			);
-			return returnData as NodeParameterValue[] | INodeParameters[];
-		}
-
-		if (parameterValue === null || parameterValue === undefined) {
-			return parameterValue;
-		}
-
-		if (typeof parameterValue !== 'object') {
-			return {};
-		}
-
-		// Data is an object
-		const returnData: INodeParameters = {};
-
-		for (const [key, value] of Object.entries(parameterValue)) {
-			returnData[key] = this.getParameterValue(
-				value as NodeParameterValueType,
-				data,
-				returnObjectAsString,
-			);
-		}
-
-		if (returnObjectAsString && typeof returnData === 'object') {
-			return this.convertObjectValueToString(returnData);
-		}
-
-		return returnData;
-	}
 }
